No need to tamper with batch size in executeBlocksWithSilkworm

+++ turbo/silkworm/call.go
@@ -25,7 +25,7 @@ import (
 	"github.com/ledgerwatch/turbo-geth/ethdb"
 )
 
-func ExecuteBlocks(funcPtr unsafe.Pointer, txn ethdb.Tx, chainID *big.Int, startBlock uint64, maxBlock uint64, batchSize uint64, writeReceipts bool) (executedBlock uint64, err error) {
+func ExecuteBlocks(funcPtr unsafe.Pointer, txn ethdb.Tx, chainID *big.Int, startBlock uint64, maxBlock uint64, batchSize int, writeReceipts bool) (executedBlock uint64, err error) {
 	cChainId := C.uint64_t(chainID.Uint64())
 	cStartBlock := C.uint64_t(startBlock)
 	cMaxBlock := C.uint64_t(maxBlock)
